refactor: extract guardian callback decision strategy from approval interception
